adapting changes after moving to Admin

+++ lms-admin/src/main/java/com/swe/lms/admin/api/service/IBookService.java
@@ -10,7 +10,7 @@ public interface IBookService {
 
     ResponseEntity<?> updateBook(BookDTO bookDTO);
 
-    ResponseEntity<?> deleteBook(int bookId);
+    ResponseEntity<?> deleteBook(long bookId);
 
     ResponseEntity<?> getBook(int bookId);
